Clippy cleanup
Fix the new unnecessary_cast lint everywhere it occurs

+++ bfffs-core/src/fs.rs
@@ -309,7 +309,7 @@ impl Uio {
 
     /// Across how many records will this UIO be spread?
     fn nrecs(&self, offset0: usize, rs: usize) -> usize {
-        (div_roundup(offset0 + self.len(), rs) - (offset0 / rs)) as usize
+        div_roundup(offset0 + self.len(), rs) - (offset0 / rs)
     }
 }
